+++ Tests/Fixtures/stream_writer/bool.php
@@ -3,7 +3,7 @@
 /**
  * @param bool $data
  */
-return static function (mixed $data, \Psr\Container\ContainerInterface $valueTransformers, array $options): \Traversable {
+return static function (mixed $data, \Psr\Container\ContainerInterface $transformers, array $options): \Traversable {
     try {
         yield $data ? 'true' : 'false';
     } catch (\JsonException $e) {