+++ Tests/Fixtures/stream_writer/null.php
@@ -3,7 +3,7 @@
 /**
  * @param null $data
  */
-return static function (mixed $data, \Psr\Container\ContainerInterface $valueTransformers, array $options): \Traversable {
+return static function (mixed $data, \Psr\Container\ContainerInterface $transformers, array $options): \Traversable {
     try {
         yield "null";
     } catch (\JsonException $e) {