+++ Tests/Fixtures/stream_writer/scalar.php
@@ -3,7 +3,7 @@
 /**
  * @param int $data
  */
-return static function (mixed $data, \Psr\Container\ContainerInterface $valueTransformers, array $options): \Traversable {
+return static function (mixed $data, \Psr\Container\ContainerInterface $transformers, array $options): \Traversable {
     try {
         yield \json_encode($data, \JSON_THROW_ON_ERROR, 512);
     } catch (\JsonException $e) {