+++ Tests/JsonStreamReaderTest.php
@@ -23,9 +23,12 @@ use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithNameAttributes;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithNullableProperties;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithPhpDoc;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithSyntheticProperties;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithValueObjects;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithValueTransformerAttributes;
-use Symfony\Component\JsonStreamer\Tests\Fixtures\ValueTransformer\DivideStringAndCastToIntValueTransformer;
-use Symfony\Component\JsonStreamer\Tests\Fixtures\ValueTransformer\StringToBooleanValueTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Transformer\DivideStringAndCastToIntValueTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Transformer\HeightValueObjectTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Transformer\StringToBooleanValueTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\ValueObject\Height;
 use Symfony\Component\TypeInfo\Type;
 use Symfony\Component\TypeInfo\TypeIdentifier;
 
@@ -120,13 +123,10 @@ class JsonStreamReaderTest extends TestCase
 
     public function testReadObjectWithValueTransformer()
     {
-        $reader = JsonStreamReader::create(
-            [
-                StringToBooleanValueTransformer::class => new StringToBooleanValueTransformer(),
-                DivideStringAndCastToIntValueTransformer::class => new DivideStringAndCastToIntValueTransformer(),
-            ],
-            $this->streamReadersDir,
-        );
+        $reader = JsonStreamReader::create([
+            StringToBooleanValueTransformer::class => new StringToBooleanValueTransformer(),
+            DivideStringAndCastToIntValueTransformer::class => new DivideStringAndCastToIntValueTransformer(),
+        ], $this->streamReadersDir);
 
         $this->assertRead($reader, function (mixed $read) {
             $this->assertInstanceOf(DummyWithValueTransformerAttributes::class, $read);
@@ -135,6 +135,27 @@ class JsonStreamReaderTest extends TestCase
             $this->assertSame('LOWERCASE NAME', $read->name);
             $this->assertSame([0, 1], $read->range);
         }, '{"id": "20", "active": "true", "name": "lowercase name", "range": "0..1"}', Type::object(DummyWithValueTransformerAttributes::class), ['scale' => 1]);
+    }
+
+    public function testReadObjectWithValueObjects()
+    {
+        $reader = JsonStreamReader::create([
+            Height::class => new HeightValueObjectTransformer(),
+        ], $this->streamReadersDir);
+
+        $this->assertRead($reader, function (mixed $read) {
+            $this->assertInstanceOf(DummyWithValueObjects::class, $read);
+            $this->assertEquals(new Height(10, 'm'), $read->height);
+            $this->assertEquals(new Height(10, 'dm'), $read->nullableHeight);
+            $this->assertEquals(new Height(10, 'cm'), $read->unionHeight);
+        }, '{"height":"10 m","nullableHeight":"10 dm","unionHeight":"10 cm"}', Type::object(DummyWithValueObjects::class));
+
+        $this->assertRead($reader, function (mixed $read) {
+            $this->assertInstanceOf(DummyWithValueObjects::class, $read);
+            $this->assertEquals(new Height(10, 'm'), $read->height);
+            $this->assertNull($read->nullableHeight);
+            $this->assertSame(10, $read->unionHeight);
+        }, '{"height":"10 m","nullableHeight":null,"unionHeight":10}', Type::object(DummyWithValueObjects::class));
     }
 
     public function testReadObjectWithPhpDoc()