+++ Tests/JsonStreamWriterTest.php
@@ -34,15 +34,19 @@ use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithPhpDoc;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithSelfReferencingDummy;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithSyntheticProperties;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithUnionProperties;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithValueObjects;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithValueTransformerAttributes;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\SelfReferencingDummy;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\SelfReferencingDummyDict;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\SelfReferencingDummyList;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\SelfReferencingDummyWithOtherDummy;
-use Symfony\Component\JsonStreamer\Tests\Fixtures\ValueTransformer\BooleanToStringValueTransformer;
-use Symfony\Component\JsonStreamer\Tests\Fixtures\ValueTransformer\DoubleIntAndCastToStringValueTransformer;
-use Symfony\Component\JsonStreamer\ValueTransformer\DateTimeToStringValueTransformer;
-use Symfony\Component\JsonStreamer\ValueTransformer\ValueTransformerInterface;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Transformer\BooleanToStringValueTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Transformer\DoubleIntAndCastToStringValueTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Transformer\HeightValueObjectTransformer;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\ValueObject\Height;
+use Symfony\Component\JsonStreamer\Transformer\DateTimeValueObjectTransformer;
+use Symfony\Component\JsonStreamer\Transformer\PropertyValueTransformerInterface;
+use Symfony\Component\JsonStreamer\Transformer\ValueObjectTransformerInterface;
 use Symfony\Component\TypeInfo\Type;
 
 class JsonStreamWriterTest extends TestCase
@@ -224,7 +228,7 @@ class JsonStreamWriterTest extends TestCase
             $dummy,
             Type::object(DummyWithValueTransformerAttributes::class),
             options: ['scale' => 1],
-            valueTransformers: [
+            transformers: [
                 BooleanToStringValueTransformer::class => new BooleanToStringValueTransformer(),
                 DoubleIntAndCastToStringValueTransformer::class => new DoubleIntAndCastToStringValueTransformer(),
             ],
@@ -242,8 +246,8 @@ class JsonStreamWriterTest extends TestCase
             $dummy,
             Type::object(DummyWithValueTransformerAttributes::class),
             options: ['scale' => 1],
-            valueTransformers: [
-                BooleanToStringValueTransformer::class => new class($this) implements ValueTransformerInterface {
+            transformers: [
+                BooleanToStringValueTransformer::class => new class($this) implements PropertyValueTransformerInterface {
                     public function __construct(
                         private JsonStreamWriterTest $test,
                     ) {
@@ -263,6 +267,35 @@ class JsonStreamWriterTest extends TestCase
                     }
                 },
                 DoubleIntAndCastToStringValueTransformer::class => new DoubleIntAndCastToStringValueTransformer(),
+            ],
+        );
+    }
+
+    public function testWriteObjectWithValueObject()
+    {
+        $dummy = new DummyWithValueObjects();
+        $dummy->height = new Height(10, 'm');
+        $dummy->nullableHeight = new Height(10, 'dm');
+        $dummy->unionHeight = new Height(10, 'cm');
+
+        $this->assertWritten(
+            '{"height":"10 m","nullableHeight":"10 dm","unionHeight":"10 cm"}',
+            $dummy,
+            Type::object(DummyWithValueObjects::class),
+            transformers: [
+                Height::class => new HeightValueObjectTransformer(),
+            ],
+        );
+
+        $dummy->nullableHeight = null;
+        $dummy->unionHeight = 10;
+
+        $this->assertWritten(
+            '{"height":"10 m","unionHeight":10}',
+            $dummy,
+            Type::object(DummyWithValueObjects::class),
+            transformers: [
+                Height::class => new HeightValueObjectTransformer(),
             ],
         );
     }
@@ -304,7 +337,7 @@ class JsonStreamWriterTest extends TestCase
             '{"interface":"2024-11-20","immutable":"2025-11-20","union":"2026-11-20"}',
             $dummy,
             Type::object(DummyWithDateTimes::class),
-            options: [DateTimeToStringValueTransformer::FORMAT_KEY => 'Y-m-d'],
+            options: [DateTimeValueObjectTransformer::FORMAT_KEY => 'Y-m-d'],
         );
 
         $dummy->union = 10;
@@ -313,7 +346,7 @@ class JsonStreamWriterTest extends TestCase
             '{"interface":"2024-11-20","immutable":"2025-11-20","union":10}',
             $dummy,
             Type::object(DummyWithDateTimes::class),
-            options: [DateTimeToStringValueTransformer::FORMAT_KEY => 'Y-m-d'],
+            options: [DateTimeValueObjectTransformer::FORMAT_KEY => 'Y-m-d'],
         );
     }
 
@@ -468,12 +501,12 @@ class JsonStreamWriterTest extends TestCase
     }
 
     /**
-     * @param array<string, mixed>                     $options
-     * @param array<string, ValueTransformerInterface> $valueTransformers
+     * @param array<string, mixed>                                                             $options
+     * @param array<string, PropertyValueTransformerInterface|ValueObjectTransformerInterface> $transformers
      */
-    private function assertWritten(string $json, mixed $data, Type $type, array $options = [], array $valueTransformers = []): void
+    private function assertWritten(string $json, mixed $data, Type $type, array $options = [], array $transformers = []): void
     {
-        $writer = JsonStreamWriter::create(streamWritersDir: $this->streamWritersDir, valueTransformers: $valueTransformers);
+        $writer = JsonStreamWriter::create($transformers, $this->streamWritersDir);
         $this->assertSame($json, (string) $writer->write($data, $type, $options));
     }
 }